+++ lib/globals.dart
@@ -3,34 +3,41 @@ library my_prj.globals;
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:tflite_flutter/tflite_flutter.dart';
-import 'package:tutorial_coach_mark/tutorial_coach_mark.dart';
 
+import 'Styling.dart';
 import 'Settings.dart';
 import 'ShowcaseTuple.dart';
 
 
 // the title of the app
-const String APP_TITLE = "Da Kanji";
+const String APP_TITLE = "DaKanji";
 
 // the whole changelog of the app
+// ignore: non_constant_identifier_names
 String WHOLE_CHANGELOG = "";
 // changelog of the newest version
+// ignore: non_constant_identifier_names
 String NEW_CHANGELOG = "";
 // should the changelog be shown
+// ignore: non_constant_identifier_names
 bool SHOW_CHANGELOG = false;
 
 // about page
+// ignore: non_constant_identifier_names
 String ABOUT = "";
 
 // the version number of this app
 // ignore: non_constant_identifier_names
 String VERSION;
 // all versions which implemented new features for the drawing screen
+// ignore: non_constant_identifier_names
 List<String> DRAWING_SCREEN_NEW_FEATURES = ["1.0.0", "1.1.0"];
 
 // the saved settings
 // ignore: non_constant_identifier_names
 final Settings SETTINGS = new Settings();
+// ignore: non_constant_identifier_names
+final Styling CURRENT_STYLING = Styling();
 
 // the global key for the drawer
 // ignore: non_constant_identifier_names
@@ -38,49 +45,9 @@ final GlobalKey<ScaffoldState> DRAWER_KEY = GlobalKey();
 
 // showcase view keys
 // ignore: non_constant_identifier_names
-final List<ShowcaseTuple> SHOWCASE_DRAWING = [
-  // 0 - Drawing canvas
-  ShowcaseTuple(GlobalKey(), "Drawing",
-    "Draw a character here", ContentAlign.bottom),
-  // 1 - undo last stroke
-  ShowcaseTuple(GlobalKey(), "Undo", 
-    "Press to undo the last stroke", ContentAlign.bottom), 
-  // 2 - clear canvas
-  ShowcaseTuple(GlobalKey(), "Clear", 
-    "Erase all strokes", ContentAlign.bottom), 
-  // 3 - prediction buttons
-  ShowcaseTuple(GlobalKey(), "Predictions", 
-    "The predicted characters will be shown here", ContentAlign.top), 
-  // 4 - short press prediction button
-  ShowcaseTuple(GlobalKey(), "Short Press Prediction",
-    "A short press copies the prediction", ContentAlign.bottom), 
-  // 5 - long press prediction button
-  ShowcaseTuple(GlobalKey(), "Long Press Prediction",
-    "A long press opens the prediction in a dictionary", ContentAlign.bottom), 
-  // 6 - multi search box
-  ShowcaseTuple(GlobalKey(), "Multi search",
-    "Here you can search multiple characters at once", ContentAlign.bottom), 
-  // 7 - double tap prediction button
-  ShowcaseTuple(GlobalKey(), "Double Tap Prediction",
-    "A Double Tap adds the character to the search box", ContentAlign.bottom), 
-  // 8 - multi search short press 
-  ShowcaseTuple(GlobalKey(), "Multi search short press",
-    "A short press copies the characters to the clipboard", ContentAlign.bottom), 
-  // 9 - multi search long press
-  ShowcaseTuple(GlobalKey(), "Multi search long press",
-    "A long press opens the characters in a dictionary", ContentAlign.bottom), 
-  // 10 - multi search double tap
-  ShowcaseTuple(GlobalKey(), "Multi search double tap",
-    "A double tap empties the field", ContentAlign.bottom), 
-  // 11 - multi search swipe left
-  ShowcaseTuple(GlobalKey(), "Multi search swipe left",
-    "Swiping left on this field deletes the last character", ContentAlign.bottom), 
-  // 12 - change dict in settings
-  ShowcaseTuple(GlobalKey(), "Dictionary Settings",
-    "In the settings the translation service can be chosen", ContentAlign.bottom), 
-];
+List<ShowcaseTuple> SHOWCASE_DRAWING = [];
+// should the showcase of the draw screen be shown
 // ignore: non_constant_identifier_names
-final Color SHOWCASE_VIGNETTE_COLOR = Color.fromARGB(255, 10, 10, 10);
 bool SHOW_SHOWCASE_DRAWING = false;
 
 // Assets
@@ -96,17 +63,17 @@ Interpreter CNN_KANJI_ONLY_INTERPRETER;
 String USED_BACKEND = "";
 
 //about page
-const GITHUB_DESKTOP_REPO = "https://github.com/CaptainDario/DaKanjiRecognizer-Desktop";
-const GITHUB_MOBILE_REPO = "https://github.com/CaptainDario/DaKanjiRecognizer-Mobile";
-const GITHUB_ML_REPO = "https://github.com/CaptainDario/DaKanjiRecognizer-ML";
-const GITHUB_ISSUES = "https://github.com/CaptainDario/DaKanjiRecognizer-Mobile/issues/new";
+const GITHUB_DESKTOP_REPO = "https://github.com/CaptainDario/DaKanji-Desktop";
+const GITHUB_MOBILE_REPO = "https://github.com/CaptainDario/Da_Kanji-Mobile";
+const GITHUB_ML_REPO = "https://github.com/CaptainDario/DaKanji-ML";
+const GITHUB_ISSUES = "https://github.com/CaptainDario/Da_Kanji-Mobile/issues/new";
 
 
 const PLAYSTORE_BASE_URL = "https://play.google.com/store/apps/details?id=";
 const PLAYSTORE_BASE_INTENT =  "market://details?id=";
 
 const APPSTORE_PAGE = "";
-const PLAYSTORE_PAGE = "https://play.google.com/store/apps/details?id=com.DaAppLab.DaKanjiRecognizer";
+const PLAYSTORE_PAGE = "https://play.google.com/store/apps/details?id=com.DaAppLab.DaKanji";
 
 const DAAPPLAB_PLAYSTORE_PAGE = "https://play.google.com/store/apps/developer?id=DaAppLab";
 const DAAPPLAB_APPSTORE_PAGE = "";