+++ lib/main.dart
@@ -1,9 +1,8 @@
-import 'package:da_kanji_recognizer_mobile/ChangelogScreen.dart';
-import 'package:da_kanji_recognizer_mobile/DrawScreen.dart';
+import 'ChangelogScreen.dart';
+import 'DrawScreen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'dart:io' show Platform;
-import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:flutter_phoenix/flutter_phoenix.dart';
 import 'package:tflite_flutter/tflite_flutter.dart';
 import 'package:package_info/package_info.dart';
@@ -16,18 +15,26 @@ import 'initInterpreter.dart';
 
 
 Future<void> main() async {
-  WidgetsFlutterBinding.ensureInitialized();
 
+  // initialize the app
+  WidgetsFlutterBinding.ensureInitialized();
   await init();
 
   runApp(
     Phoenix(
-      child: DaKanjiRecognizerApp(),
+      child: DaKanjiApp(),
     )
   );
 
 }
 
+
+/// Initializes the app.
+/// 
+/// This function initializes:
+/// * reads used version, CHANGELOG about from file
+/// * loads the settings
+/// * initializes tensorflow lite and reads the labels from file 
 Future<void> init() async {
   // get the app's version 
   VERSION = (await PackageInfo.fromPlatform()).version;
@@ -40,9 +47,9 @@ Future<void> init() async {
     CNN_KANJI_ONLY_INTERPRETER = await initInterpreterAndroid();
   else if (Platform.isIOS) 
     CNN_KANJI_ONLY_INTERPRETER = await initInterpreterIOS();
-  else if (kIsWeb) 
-    CNN_KANJI_ONLY_INTERPRETER = await initInterpreterWeb();
-  
+  else
+    throw PlatformException(code: "Platform not supported.");
+
   // load labels, CHANGELOG and about from file
   String labels = 
     await rootBundle.loadString("assets/labels_CNN_kanji_only.txt");
@@ -63,13 +70,17 @@ Future<void> init() async {
   CNN_KANJI_ONLY_INTERPRETER.run(_input, _output);
 }
 
+/// Reads `CHANGELOG.md` from file and returns a converted version.
+/// 
+/// First reads the changelog from file and than returns a list with the changes 
+/// in the current version and the whole changelog.
 Future<List<String>> initChangelog () async {
 
-  String changelog = await rootBundle.loadString("assets/CHANGELOG.md");
+  String changelog = await rootBundle.loadString("CHANGELOG.md");
   // whole changelog
-  List<String> changelog_list = changelog.split("\n");
-  changelog_list.removeRange(0, 3);
-  String wholeChangelog = changelog_list.join("\n");
+  List<String> changelogList = changelog.split("\n");
+  changelogList.removeRange(0, 3);
+  String wholeChangelog = changelogList.join("\n");
   // newest changes
   final matches = new RegExp(r"(##.*?##)", dotAll: true);
   String newestChangelog = matches.firstMatch(changelog).group(0).toString();
@@ -78,6 +89,7 @@ Future<List<String>> initChangelog () async {
   return [newestChangelog, wholeChangelog];
 }
 
+/// Reads `about.md` from file and returns a converted version.
 Future<String> initAbout () async {
 
   String about = await rootBundle.loadString("assets/about.md");
@@ -105,7 +117,8 @@ Future<String> initAbout () async {
 }
 
 
-class DaKanjiRecognizerApp extends StatelessWidget {
+/// The starting widget of the app
+class DaKanjiApp extends StatelessWidget {
   
   @override
   Widget build(BuildContext context) {
@@ -129,7 +142,8 @@ class DaKanjiRecognizerApp extends StatelessWidget {
       themeMode: SETTINGS.themesDict[SETTINGS.selectedTheme],
 
       //screens
-      home: HomeScreen(),
+      initialRoute: "/home",
+      //home: HomeScreen(),
       routes: <String, WidgetBuilder>{
         "/home": (BuildContext context) => HomeScreen(),
         "/drawing": (BuildContext context) => DrawScreen(),