+++ CHANGELOG.md
@@ -12,11 +12,12 @@ New Features:
 - What's new message
 
 Changes:
-- renamed to Da Kanji
+- beautiful new icon
+- renamed to DaKanji
 - Refactoring
+- Provide apk on GitHub
 
 Fixes:
-- stuck at loading screen on some devices with API <= 22
 - "character copied" message on some devices not showing up 
 - UI layout on certain aspect ratios
 - show tutorial **only** if there are new features